streamlit_app/requirements.txt
requests
streamlit==1.28.2
python-pptx==0.6.23
pypdf==3.17.1
python-docx==1.1.0
kss==4.5.4
mecab-python==1.0.0
python-dotenv==1.0.0
